+++ tomcat9/src/test/java/com/hazelcast/session/springboot/AbstractSpringBootConfigurationTest.java
@@ -28,18 +28,31 @@ public abstract class AbstractSpringBootConfigurationTest {
         assertNotNull(applicationContext.getBean("hazelcastTomcatSessionManagerCustomizer"));
     }
 
-    @Test()
-    public void testSessionManager()
-            throws Exception {
+    @Test
+    public void testSessionMapCreated() {
+        //given
+        //the Spring Boot application is started
+        //then
         HazelcastInstance hazelcastInstance = (HazelcastInstance) applicationContext.getBean("hazelcastInstance");
         LinkedList<DistributedObject> distributedObjects = (LinkedList<DistributedObject>) hazelcastInstance.getDistributedObjects();
         assertEquals("Session map should be created.", 1, distributedObjects.size());
+    }
 
+    @Test()
+    public void testSessionStoredInHazelcast()
+            throws Exception {
+        //given
+        //the Spring Boot application is started
+        HazelcastInstance hazelcastInstance = (HazelcastInstance) applicationContext.getBean("hazelcastInstance");
+        LinkedList<DistributedObject> distributedObjects = (LinkedList<DistributedObject>) hazelcastInstance.getDistributedObjects();
+
+        //when
         CookieStore cookieStore = new BasicCookieStore();
         HttpClient client = HttpClientBuilder.create().setDefaultCookieStore(cookieStore).build();
         HttpGet request = new HttpGet("http://localhost:9999/set");
         client.execute(request);
 
+        //then
         IMap<Object, Object> sessionMap = hazelcastInstance.getMap(distributedObjects.get(0).getName());
         assertEquals("Session should be created.",1, sessionMap.size());
     }
