--- tomcat9/src/test/java/com/hazelcast/session/springboot/ClientServerApplication.java
@@ -1,13 +1,0 @@
-package com.hazelcast.session.springboot;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.PropertySource;
-
-@SpringBootApplication
-@PropertySource("classpath:clientServer.properties")
-public class ClientServerApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(ClientServerApplication.class, args);
-    }
-}
