+++ tomcat9/src/test/java/com/hazelcast/session/springboot/ClientServerSpringBootConfigurationTest.java
@@ -5,6 +5,8 @@ import com.hazelcast.core.HazelcastInstance;
 import org.junit.After;
 import org.junit.Before;
 import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.PropertySource;
 
 public class ClientServerSpringBootConfigurationTest
         extends AbstractSpringBootConfigurationTest {
@@ -21,5 +23,13 @@ public class ClientServerSpringBootConfigurationTest
     public void clean() {
         SpringApplication.exit(applicationContext);
         hazelcastInstance.shutdown();
+    }
+
+    @SpringBootApplication
+    @PropertySource("classpath:clientServer.properties")
+    static class ClientServerApplication {
+        public static void main(String[] args) {
+            SpringApplication.run(ClientServerApplication.class, args);
+        }
     }
 }
