+++ tomcat9/src/test/java/com/hazelcast/session/springboot/P2PSpringBootConfigurationTest.java
@@ -3,6 +3,9 @@ package com.hazelcast.session.springboot;
 import org.junit.After;
 import org.junit.Before;
 import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.FilterType;
 
 public class P2PSpringBootConfigurationTest
         extends AbstractSpringBootConfigurationTest {
@@ -17,4 +20,12 @@ public class P2PSpringBootConfigurationTest
         SpringApplication.exit(applicationContext);
     }
 
+    @SpringBootApplication
+    @ComponentScan(excludeFilters = @ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE,
+            classes = ClientServerSpringBootConfigurationTest.ClientServerApplication.class))
+    static class P2PApplication {
+        public static void main(String[] args) {
+            SpringApplication.run(P2PApplication.class, args);
+        }
+    }
 }
